Refactor patient data fetching logic into a separate module for improved readability and reusability
